requirements.txt
# backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
google-generativeai
langchain==0.0.340
langgraph==0.0.20
pypdf2==3.0.1
pydantic==2.4.2
python-dotenv==1.0.0

# frontend
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
requests==2.31.0

# utils
tiktoken==0.5.1
chromadb==0.4.18